Major refactor, introduce modules for separation of s3 & resize functions
- Remove hard-coding of output image name
- Add info logging
- Minor changes
