+++ rs-serverless-resizing/Cargo.toml
@@ -8,10 +8,6 @@ edition = "2018"
 
  
 [[bin]]
-name = "resize_rust"
-path = "src/img_resizing_lambda.rs"
-
-[[bin]]
 name = "lambda"
 path = "src/lambda_func.rs"
